Refactored dashboard component
