+++ fathom/src/components/Dashboard/Dashboard.tsx
@@ -1,77 +1,21 @@
 import * as React from 'react';
-import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
-import CssBaseline from '@mui/material/CssBaseline';
+import { styled, createTheme } from '@mui/material/styles';
 import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
 
 import Toolbar from '@mui/material/Toolbar';
-import List from '@mui/material/List';
-import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
-import IconButton from '@mui/material/IconButton';
-// import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
-import MenuIcon from '@mui/icons-material/Menu';
-import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-// import NotificationsIcon from '@mui/icons-material/Notifications';
-import AccountBalanceWalletIcon from '@mui/icons-material/AccountBalanceWallet';
-import LogoutIcon from '@mui/icons-material/Logout';
-import Copyright from '../Footer/Footer';
-import AppBar from '../AppBar/AppBar';
-import { mainListItems, secondaryListItems } from './listItems';
 import useMetaMask from '../../hooks/metamask';
-import { Chip } from '@mui/material';
 import PoolListView from '../Pools/PoolListView';
 import OpenPositionsList from '../PositionList/OpenPositionsList';
-import StableSwap from '../Stableswap/StableSwap';
-import { Route, Routes } from 'react-router-dom';
 import { useStores } from '../../stores';
 import { LogLevel, useLogger } from '../../helpers/Logger';
 import { useEffect } from 'react';
 import { observer } from 'mobx-react';
 
-// import Chart from './Chart';
-// import Deposits from './Deposits';
-// import Orders from './Orders';
-
-
-
-const drawerWidth: number = 240;
-
-const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
-  ({ theme, open }) => ({
-    '& .MuiDrawer-paper': {
-      position: 'relative',
-      whiteSpace: 'nowrap',
-      width: drawerWidth,
-      transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-      }),
-      boxSizing: 'border-box',
-      ...(!open && {
-        overflowX: 'hidden',
-        transition: theme.transitions.create('width', {
-          easing: theme.transitions.easing.sharp,
-          duration: theme.transitions.duration.leavingScreen,
-        }),
-        width: theme.spacing(7),
-        [theme.breakpoints.up('sm')]: {
-          width: theme.spacing(9),
-        },
-      }),
-    },
-  }),
-);
-
-const mdTheme = createTheme();
-
 const DashboardContent = observer(() => {
-
-  const [open, setOpen] = React.useState(true);
-  const { connect, isActive, account, disconnect } = useMetaMask()!
 
   let poolStore = useStores().poolStore;
   let logger = useLogger();
@@ -82,67 +26,7 @@ const DashboardContent = observer(() => {
     poolStore.fetchPools()
   },[poolStore]);
 
-  const toggleDrawer = () => {
-    setOpen(!open);
-  };
-
   return (
-    <ThemeProvider theme={mdTheme}>
-      <Box sx={{ display: 'flex' }}>
-        <CssBaseline />
-        <AppBar position="absolute" open={open}>
-          <Toolbar
-            sx={{
-              pr: '24px', // keep right padding when drawer closed
-            }}
-          >
-            <IconButton
-              edge="start"
-              color="inherit"
-              aria-label="open drawer"
-              onClick={toggleDrawer}
-              sx={{
-                marginRight: '36px',
-                ...(open && { display: 'none' }),
-              }}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography
-              component="h1"
-              variant="h6"
-              color="inherit"
-              noWrap
-              sx={{ flexGrow: 1 }}
-            >
-              Fathom Stablecoin
-            </Typography>
-            <Chip label={account} color="primary" />
-            <IconButton color="inherit" >
-                {isActive  ?  <LogoutIcon /> : <AccountBalanceWalletIcon onClick={connect}/>}
-            </IconButton>
-          </Toolbar>
-        </AppBar>
-        <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              px: [1],
-            }}
-          >
-            <IconButton onClick={toggleDrawer}>
-              <ChevronLeftIcon />
-            </IconButton>
-          </Toolbar>
-          <Divider />
-          <List component="nav">
-            {mainListItems}
-            <Divider sx={{ my: 1 }} />
-            {secondaryListItems}
-          </List>
-        </Drawer>
         <Box
           component="main"
           sx={{
@@ -179,20 +63,12 @@ const DashboardContent = observer(() => {
                   </Grid>
              )}
               <Grid item xs={12}>
-                <Routes>
-                  <Route path="/" element={<OpenPositionsList />} />
-                  <Route path="/swap" element={<StableSwap />} />
-                </Routes>
+                <OpenPositionsList />
               </Grid>
             </Grid>
-            <Copyright sx={{ pt: 4 }} />
           </Container>
         </Box>
-      </Box>
-    </ThemeProvider>
   );
 })
 
 export default DashboardContent;
-
-//export default DashboardContent;
